move test to esql module
